refactor(repl): extract pager_ui and state/sponge modules
Move the IPC alternate-screen pager helpers out of repl/mod.rs and
SpongeQueue out of repl/state.rs into dedicated modules. Pure code
movement, no behavior change.
